[Rename] 폴더 생성 후 해당 파일들 경로 변경
